Several optimizations...

+++ src/graphic/surface.cpp
@@ -309,7 +309,7 @@ void Surface::MergeSurface( Surface &spr, const Point2i &pos)
 {
   Uint32 spr_pix, cur_pix;
   Uint8 r, g, b, a, p_r, p_g, p_b, p_a;
-  double f_a, f_ca, f_pa;
+  float f_a, f_ca, f_pa;
   SDL_PixelFormat * current_fmt = surface->format;
   SDL_PixelFormat * spr_fmt = spr.surface->format;
   int current_offset, spr_offset;
@@ -317,34 +317,43 @@ void Surface::MergeSurface( Surface &spr, const Point2i &pos)
 
   spr.Lock();
   Lock();
+
   // for each pixel lines of a source image
   for (offset.x = (pos.x > 0 ? 0 : -pos.x); offset.x < spr.GetWidth() && pos.x + offset.x < GetWidth(); offset.x++) {
     for (offset.y = (pos.y > 0 ? 0 : -pos.y); offset.y < spr.GetHeight() && pos.y + offset.y < GetHeight(); offset.y++) {
+
       // Computing offset on both sprite
       current_offset = (pos.y + offset.y) * surface->w + pos.x + offset.x;
       spr_offset = offset.y * spr.surface->w + offset.x;
+
       // Retrieving a pixel of sprite to merge
       spr_pix = ((Uint32*)spr.surface->pixels)[spr_offset];
       cur_pix = ((Uint32*)surface->pixels)[current_offset];
+
       // Retreiving each chanel of the pixel using pixel format
       r = (Uint8)(((spr_pix & spr_fmt->Rmask) >> spr_fmt->Rshift) << spr_fmt->Rloss);
       g = (Uint8)(((spr_pix & spr_fmt->Gmask) >> spr_fmt->Gshift) << spr_fmt->Gloss);
       b = (Uint8)(((spr_pix & spr_fmt->Bmask) >> spr_fmt->Bshift) << spr_fmt->Bloss);
       a = (Uint8)(((spr_pix & spr_fmt->Amask) >> spr_fmt->Ashift) << spr_fmt->Aloss);
+
       // Retreiving previous alpha value
       p_a = (Uint8)(((cur_pix & current_fmt->Amask) >> current_fmt->Ashift) << current_fmt->Aloss);
-      if(a == SDL_ALPHA_OPAQUE || (p_a == 0 && a >0)) // new pixel with no alpha or nothing on previous pixel
+
+      if (a == SDL_ALPHA_OPAQUE || (p_a == 0 && a >0)) // new pixel with no alpha or nothing on previous pixel
         ((Uint32 *)(surface->pixels))[current_offset] = SDL_MapRGBA(current_fmt, r, g, b, a);
       else if (a > 0) { // alpha is lower => merge color with previous value
-        f_a = (double)a / 255.0;
+        f_a = (float)a / 255.0;
         f_ca = 1.0 - f_a;
-        f_pa = (double)p_a / 255.0;
+        f_pa = (float)p_a / 255.0;
+
         p_r = (Uint8)(((cur_pix & current_fmt->Rmask) >> current_fmt->Rshift) << current_fmt->Rloss);
         p_g = (Uint8)(((cur_pix & current_fmt->Gmask) >> current_fmt->Gshift) << current_fmt->Gloss);
         p_b = (Uint8)(((cur_pix & current_fmt->Bmask) >> current_fmt->Bshift) << current_fmt->Bloss);
-        r = (Uint8)((double)p_r * f_ca * f_pa + (double)r * f_a);
-        g = (Uint8)((double)p_g * f_ca * f_pa + (double)g * f_a);
-        b = (Uint8)((double)p_b * f_ca * f_pa + (double)b * f_a);
+
+        r = (Uint8)((float)p_r * f_ca * f_pa + (float)r * f_a);
+        g = (Uint8)((float)p_g * f_ca * f_pa + (float)g * f_a);
+        b = (Uint8)((float)p_b * f_ca * f_pa + (float)b * f_a);
+
         a = (a > p_a ? a : p_a);
         ((Uint32 *)(surface->pixels))[current_offset] = SDL_MapRGBA(current_fmt, r, g, b, a);
       }
